fix: compiler warnings from curl deprecations, int/double mixing, sizer flags
Warnings surfaced by the 2.0.0 Debian package build on current toolchains:

- AlpacaClient::Get(): CURLINFO_CONTENT_LENGTH_DOWNLOAD and
  CURLINFO_SIZE_DOWNLOAD are deprecated since curl 7.55 in favor of the
  curl_off_t _T variants; switch to those and log with %lld.
- backlash_comp.cpp / scope.cpp: the BACKLASH_* / BL_* / TRIAL_* distance
  constants are ints compared and scaled against doubles; add explicit
  (double) casts so -Wconversion-style warnings stay quiet and the intent
  is visible.
- calreview_dialog.cpp: wxStaticBoxSizer's ctor takes an orientation, not
  sizer-item flags — wxEXPAND in that argument trips a wxWidgets 3.2
  assert/warning; drop it (expansion is set where the sizer is added).

Full Release build now compiles with zero project warnings.

+++ src/scope.cpp
@@ -1363,7 +1363,7 @@ bool Scope::UpdateCalibrationState(const PHD_Point& currentLocation)
                 else
                 {
                     // Used up all our attempts - might be ok or not
-                    if (blCumDelta >= BL_MIN_CLEARING_DISTANCE)
+                    if (blCumDelta >= (double) BL_MIN_CLEARING_DISTANCE)
                     {
                         // Exhausted all the clearing pulses without reaching the goal - but we did move the mount > 3 px (same
                         // as PHD1)
